+++ src/fps_panel.h
@@ -7,13 +7,13 @@
 
 typedef struct {
     Panel panel;
-    double fps;
+    float fps;
 } FPSPanel;
 
 FPSPanel* fps_panel_init(FPSPanel* p, PanelManager* panelManager);
 void fps_panel_destroy(FPSPanel* fpsPanel);
 
-void fps_panel_set_fps(FPSPanel* fpsPanel, double fps);
+void fps_panel_set_fps(FPSPanel* fpsPanel, float fps);
 void fps_panel_set_position(FPSPanel* fpsPanel, unsigned int x, unsigned int y);
 
 #endif // FPS_PANEL_H
